DEVELOP-001: Fix refresh related div error
this.parentNode.id breaks only works if the slot has not refreshed by the time you go to submit your form.

This causes problems with refreshing units as there's no guarantee your slot will not have refreshed. We store the clicked div into adentify.clickedDiv instead.

If the user leaves that modal and opens up a new one in a different slot, adentify.clickedDiv is updated to the new div ID